requirements.txt
requests==2.23.0
beautifulsoup4==4.9.1
lxml==4.5.2
pandas==1.1.1
python-telegram-bot==12.8
